chore(experiments): extract flag-cleanup logic into a tested module
